talos/pyproject.toml
[project]
name = "talos"
version = "0.1.0"
description = "Smarthome build tooling"
requires-python = ">=3.10"
dependencies = [
    "click",
    "Jinja2",
    "PyYAML",
    "rich",
    "tomli",
]

[project.scripts]
talos = "talos.cli:app"

[project.optional-dependencies]
test = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/talos"]
include = ["src/talos/templates/**"]
